controller: Move manager to a dedicated file
