+++ internal/controller/kustomization_controller.go
@@ -36,14 +36,9 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	kerrors "k8s.io/apimachinery/pkg/util/errors"
 	kuberecorder "k8s.io/client-go/tools/record"
-	"k8s.io/client-go/util/workqueue"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/fluxcd/cli-utils/pkg/kstatus/polling"
@@ -64,7 +59,6 @@ import (
 	runtimeCtrl "github.com/fluxcd/pkg/runtime/controller"
 	"github.com/fluxcd/pkg/runtime/jitter"
 	"github.com/fluxcd/pkg/runtime/patch"
-	"github.com/fluxcd/pkg/runtime/predicates"
 	"github.com/fluxcd/pkg/runtime/statusreaders"
 	"github.com/fluxcd/pkg/ssa"
 	"github.com/fluxcd/pkg/ssa/normalize"
@@ -114,127 +108,6 @@ type KustomizationReconciler struct {
 	GroupChangeLog             bool
 	AdditiveCELDependencyCheck bool
 	TokenCache                 *cache.TokenCache
-}
-
-// KustomizationReconcilerOptions contains options for the KustomizationReconciler.
-type KustomizationReconcilerOptions struct {
-	HTTPRetry                 int
-	DependencyRequeueInterval time.Duration
-	RateLimiter               workqueue.TypedRateLimiter[reconcile.Request]
-	WatchConfigsPredicate     predicate.Predicate
-}
-
-func (r *KustomizationReconciler) SetupWithManager(ctx context.Context, mgr ctrl.Manager, opts KustomizationReconcilerOptions) error {
-	const (
-		indexOCIRepository = ".metadata.ociRepository"
-		indexGitRepository = ".metadata.gitRepository"
-		indexBucket        = ".metadata.bucket"
-		indexConfigMap     = ".metadata.configMap"
-		indexSecret        = ".metadata.secret"
-	)
-
-	// Index the Kustomizations by the OCIRepository references they (may) point at.
-	if err := mgr.GetCache().IndexField(ctx, &kustomizev1.Kustomization{}, indexOCIRepository,
-		r.indexBy(sourcev1.OCIRepositoryKind)); err != nil {
-		return fmt.Errorf("failed creating index %s: %w", indexOCIRepository, err)
-	}
-
-	// Index the Kustomizations by the GitRepository references they (may) point at.
-	if err := mgr.GetCache().IndexField(ctx, &kustomizev1.Kustomization{}, indexGitRepository,
-		r.indexBy(sourcev1.GitRepositoryKind)); err != nil {
-		return fmt.Errorf("failed creating index %s: %w", indexGitRepository, err)
-	}
-
-	// Index the Kustomizations by the Bucket references they (may) point at.
-	if err := mgr.GetCache().IndexField(ctx, &kustomizev1.Kustomization{}, indexBucket,
-		r.indexBy(sourcev1.BucketKind)); err != nil {
-		return fmt.Errorf("failed creating index %s: %w", indexBucket, err)
-	}
-
-	// Index the Kustomization by the ConfigMap references they point to.
-	if err := mgr.GetFieldIndexer().IndexField(ctx, &kustomizev1.Kustomization{}, indexConfigMap,
-		func(o client.Object) []string {
-			obj := o.(*kustomizev1.Kustomization)
-			namespace := obj.GetNamespace()
-			var keys []string
-			if kc := obj.Spec.KubeConfig; kc != nil && kc.ConfigMapRef != nil {
-				keys = append(keys, fmt.Sprintf("%s/%s", namespace, kc.ConfigMapRef.Name))
-			}
-			if pb := obj.Spec.PostBuild; pb != nil {
-				for _, ref := range pb.SubstituteFrom {
-					if ref.Kind == "ConfigMap" {
-						keys = append(keys, fmt.Sprintf("%s/%s", namespace, ref.Name))
-					}
-				}
-			}
-			return keys
-		},
-	); err != nil {
-		return fmt.Errorf("failed creating index %s: %w", indexConfigMap, err)
-	}
-
-	// Index the Kustomization by the Secret references they point to.
-	if err := mgr.GetFieldIndexer().IndexField(ctx, &kustomizev1.Kustomization{}, indexSecret,
-		func(o client.Object) []string {
-			obj := o.(*kustomizev1.Kustomization)
-			namespace := obj.GetNamespace()
-			var keys []string
-			if dec := obj.Spec.Decryption; dec != nil && dec.SecretRef != nil {
-				keys = append(keys, fmt.Sprintf("%s/%s", namespace, dec.SecretRef.Name))
-			}
-			if kc := obj.Spec.KubeConfig; kc != nil && kc.SecretRef != nil {
-				keys = append(keys, fmt.Sprintf("%s/%s", namespace, kc.SecretRef.Name))
-			}
-			if pb := obj.Spec.PostBuild; pb != nil {
-				for _, ref := range pb.SubstituteFrom {
-					if ref.Kind == "Secret" {
-						keys = append(keys, fmt.Sprintf("%s/%s", namespace, ref.Name))
-					}
-				}
-			}
-			return keys
-		},
-	); err != nil {
-		return fmt.Errorf("failed creating index %s: %w", indexSecret, err)
-	}
-
-	r.requeueDependency = opts.DependencyRequeueInterval
-	r.statusManager = fmt.Sprintf("gotk-%s", r.ControllerName)
-	r.artifactFetchRetries = opts.HTTPRetry
-
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&kustomizev1.Kustomization{}, builder.WithPredicates(
-			predicate.Or(predicate.GenerationChangedPredicate{}, predicates.ReconcileRequestedPredicate{}),
-		)).
-		Watches(
-			&sourcev1.OCIRepository{},
-			handler.EnqueueRequestsFromMapFunc(r.requestsForRevisionChangeOf(indexOCIRepository)),
-			builder.WithPredicates(SourceRevisionChangePredicate{}),
-		).
-		Watches(
-			&sourcev1.GitRepository{},
-			handler.EnqueueRequestsFromMapFunc(r.requestsForRevisionChangeOf(indexGitRepository)),
-			builder.WithPredicates(SourceRevisionChangePredicate{}),
-		).
-		Watches(
-			&sourcev1.Bucket{},
-			handler.EnqueueRequestsFromMapFunc(r.requestsForRevisionChangeOf(indexBucket)),
-			builder.WithPredicates(SourceRevisionChangePredicate{}),
-		).
-		WatchesMetadata(
-			&corev1.ConfigMap{},
-			handler.EnqueueRequestsFromMapFunc(r.requestsForConfigDependency(indexConfigMap)),
-			builder.WithPredicates(predicate.ResourceVersionChangedPredicate{}, opts.WatchConfigsPredicate),
-		).
-		WatchesMetadata(
-			&corev1.Secret{},
-			handler.EnqueueRequestsFromMapFunc(r.requestsForConfigDependency(indexSecret)),
-			builder.WithPredicates(predicate.ResourceVersionChangedPredicate{}, opts.WatchConfigsPredicate),
-		).
-		WithOptions(controller.Options{
-			RateLimiter: opts.RateLimiter,
-		}).
-		Complete(r)
 }
 
 func (r *KustomizationReconciler) Reconcile(ctx context.Context, req ctrl.Request) (result ctrl.Result, retErr error) {
